[FEAT] 운동 기록 디테일 페이지 #74
